+++ tomcatmanager/tomcat_manager.py
@@ -30,7 +30,7 @@ import re
 
 import requests
 
-from .models import codes, TomcatManagerResponse, ServerInfo, TomcatApplication
+from .models import status_codes, TomcatManagerResponse, ServerInfo, TomcatApplication
 
 
 class TomcatManager:
@@ -533,11 +533,11 @@ class TomcatManager:
         # we have to force a status_code and a status_message
         # because the server doesn't return them
         if r.response.status_code == requests.codes.ok:
-            r.status_code = codes.ok
-            r.status_message = codes.ok
+            r.status_code = status_codes.ok
+            r.status_message = status_codes.ok
         else:
-            r.status_code = codes.fail
-            r.status_message = codes.fail
+            r.status_code = status_codes.fail
+            r.status_message = status_codes.fail
         return r
 
     def vm_info(self):
@@ -604,7 +604,7 @@ class TomcatManager:
         resources = {}
         for line in r.result.splitlines():
             resource, classname = line.rstrip().split(':', 1)
-            if resource[:7] != codes.fail + ' - ':
+            if resource[:7] != status_codes.fail + ' - ':
                 resources[resource] = classname.lstrip()
         r.resources = resources
         return r